refactor: Phase 2 — split parser/matching into focused submodules
Replace the 1160-line matching.rs with matching/{mod,capture,nested_dict,
custom_type,raw,py_match}.rs for SRP: capture geometry, nested dict paths,
custom-type validation, raw batch matching, and Python match paths are
isolated while preserving the public matching API.
